# gen/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Greeter API
    version: 0.0.1
paths:
    /v1/greeter/hello:
        post:
            tags:
                - Greeter
            description: Unary RPC with a REST/HTTP mapping consumed by the gateway + OpenAPI.
            operationId: Greeter_SayHello
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/HelloRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HelloReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
                "400":
                    description: One or more validation errors occurred.
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ValidationProblemDetails'
    /v1/greeter/{name}:
        get:
            tags:
                - Greeter
            operationId: Greeter_GetGreeting
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HelloReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
                "400":
                    description: One or more validation errors occurred.
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ValidationProblemDetails'
components:
    schemas:
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        HelloReply:
            type: object
            properties:
                message:
                    type: string
        HelloRequest:
            type: object
            properties:
                name:
                    type: string
                    description: 'protovalidate constraints: enforced at runtime, no codegen needed.'
                    minLength: 1
                    maxLength: 40
                times:
                    type: integer
                    format: int32
                    minimum: 1
                    maximum: 10
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        ValidationProblemDetails:
            type: object
            description: RFC 9457 problem+json returned when request validation fails.
            properties:
                type:
                    type: string
                title:
                    type: string
                status:
                    type: integer
                    format: int32
                errors:
                    type: object
                    description: Map of field name (JSON) to its validation messages.
                    additionalProperties:
                        type: array
                        items:
                            type: string
tags:
    - name: Greeter